File made for Carly
Jon_and_carly_lock_in_sim is the main burrito here and what should be worked on. The OG lock in software was made before V1 of the firmware but still has useful functions to reference. The third file mag_then_mix was the software simulation of V2 of the firmware where magnitudes were taken directly after the FFT and before the demodulation mix. In the newest version, we are going back to my original plan (yay) and demodulating directly after the FFT and before taking magnitudes